chore(loading): integrate loading components
